+++ l10n_br_fiscal/models/document_fiscal_line_mixin.py
@@ -551,17 +551,15 @@ class FiscalDocumentLineMixin(models.AbstractModel):
         domain=[("tax_domain", "=", TAX_DOMAIN_II)],
     )
 
-    ii_base = fields.Float(string="II Base", digits=dp.get_precision("Account"))
+    ii_base = fields.Monetary(string="II Base")
 
     ii_percent = fields.Float(string="II %")
 
-    ii_value = fields.Float(string="II Value", digits=dp.get_precision("Account"))
+    ii_value = fields.Monetary(string="II Value")
 
-    ii_iof_value = fields.Float(string="IOF Value", digits=dp.get_precision("Account"))
+    ii_iof_value = fields.Monetary(string="IOF Value")
 
-    ii_customhouse_charges = fields.Float(
-        string="Despesas Aduaneiras", digits=dp.get_precision("Account")
-    )
+    ii_customhouse_charges = fields.Monetary(string="Despesas Aduaneiras")
 
     # PIS/COFINS Fields
     # COFINS
