+++ l10n_br_fiscal/models/document_fiscal_mixin.py
@@ -176,6 +176,21 @@ class FiscalDocumentMixin(models.AbstractModel):
         compute="_compute_amount",
     )
 
+    amount_ii_base = fields.Monetary(
+        string="II Base",
+        compute="_compute_amount",
+    )
+
+    amount_ii_value = fields.Monetary(
+        string="II Value",
+        compute="_compute_amount",
+    )
+
+    amount_ii_customhouse_charges = fields.Monetary(
+        string="Customhouse Charges",
+        compute="_compute_amount",
+    )
+
     amount_pis_base = fields.Monetary(
         string="PIS Base",
         compute="_compute_amount",
