+++ l10n_br_nfe/models/document.py
@@ -16,7 +16,7 @@ from lxml import etree
 from nfelib.v4_00 import retEnviNFe as leiauteNFe
 from requests import Session
 
-from odoo import _, api, fields
+from odoo import _, api, fields, models
 from odoo.exceptions import UserError
 
 from odoo.addons.l10n_br_fiscal.constants.fiscal import (
@@ -46,20 +46,19 @@ from ..constants.nfe import NFE_ENVIRONMENTS, NFE_VERSIONS
 _logger = logging.getLogger(__name__)
 
 
-def filter_processador_edoc_nfe(record):
-    if record.processador_edoc == PROCESSADOR_OCA and record.document_type_id.code in [
-        MODELO_FISCAL_NFE,
-        MODELO_FISCAL_NFCE,
-    ]:
-        return True
-    return False
-
+class NFe(models.Model):
+    _name = "l10n_br_fiscal.document"
+    _inherit = [
+        "l10n_br_fiscal.document",
+         "l10n_br_nfe.document.workflow",
+         "l10n_br_nfe.document.electronic",
+    ]
 
 class NFe(spec_models.StackedModel):
     _name = "l10n_br_fiscal.document"
     _inherit = ["l10n_br_fiscal.document", "nfe.40.infnfe"]
     _stacked = "nfe.40.infnfe"
-    _stack_skip = "nfe40_veicTransp"
+    _stack_skip = "nfe40_veicTransp" # TODO Check se as proximas tags não estao stacked como a nfe.40.exporta por causa disso
     _field_prefix = "nfe40_"
     _schema_name = "nfe"
     _schema_version = "4.0.0"
@@ -79,9 +78,8 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: infNFe
     ##########################
-    nfe40_versao = fields.Char(
-        related="document_version",
-    )
+    
+    nfe40_versao = fields.Char(related="document_version")
 
     nfe_version = fields.Selection(
         selection=NFE_VERSIONS,
@@ -90,20 +88,25 @@ class NFe(spec_models.StackedModel):
         default=lambda self: self.env.user.company_id.nfe_version,
     )
 
-    nfe40_Id = fields.Char(
-        related="document_key",
-    )
+    nfe40_Id = fields.Char(related="document_key")
 
     ##########################
     # NF-e tag: ide
     ##########################
-    nfe40_cUF = field.Char(related="company_id.partner_id.state_id.ibge_code") # TODO criar uma função para tratar quando for entrada
+
+    nfe40_cUF = fields.Char( # TODO criar uma função para tratar quando for entrada
+        related="company_id.partner_id.state_id.ibge_code",
+        string="nfe40_cUF",
+    )
 
     # <cNF>17983659</cNF> TODO
 
     nfe40_natOp = fields.Char(related="operation_name")
 
-    nfe40_mod = field.Char(related="document_type_id.code")
+    nfe40_mod = fields.Char(
+        related="document_type_id.code",
+        string="nfe40_mod",
+    )
 
     nfe40_serie = fields.Char(related="document_serie")
 
@@ -120,7 +123,7 @@ class NFe(spec_models.StackedModel):
 
     nfe40_idDest = fields.Selection(compute="_compute_nfe40_idDest")
 
-    cMunFG = fieds.Char(related="company_id.partner_id.city_id.ibge_code")
+    cMunFG = fields.Char(related="company_id.partner_id.city_id.ibge_code")
 
     nfe40_tpImp = fields.Selection(default="1")
 
@@ -152,6 +155,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: NFref
     ##########################
+
     nfe40_NFref = fields.One2many(
         comodel_name="l10n_br_fiscal.document.related",
         related="document_related_ids",
@@ -164,6 +168,7 @@ class NFe(spec_models.StackedModel):
     # emit and dest are not related fields as their related fields
     # can change depending if it's and incoming our outgoing NFe
     # specially when importing (ERP NFe migration vs supplier Nfe).
+
     nfe40_emit = fields.Many2one(
         comodel_name="res.company", compute="_compute_emit", readonly=True, string="Emit"
     )
@@ -180,6 +185,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: dest
     ##########################
+
     nfe40_dest = fields.Many2one(
         comodel_name="res.partner", compute="_compute_dest", readonly=True, string="Dest"
     )
@@ -206,6 +212,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: det
     ##########################
+
     # TODO should be done by framework?
     nfe40_det = fields.One2many(
         comodel_name="l10n_br_fiscal.document.line",
@@ -216,6 +223,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: ICMSTot
     ##########################
+
     nfe40_vBC = fields.Monetary(string="BC do ICMS", related="amount_icms_base")
 
     nfe40_vICMS = fields.Monetary(related="amount_icms_value")
@@ -268,16 +276,19 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: transp
     ##########################
+
     nfe40_modFrete = fields.Selection(default="9")
 
     ##########################
     # NF-e tag: transporta
     ##########################
+
     nfe40_transporta = fields.Many2one(comodel_name="res.partner")
 
     ##########################
     # NF-e tag: pag
     ##########################
+
     def _prepare_amount_financial(self, ind_pag, t_pag, v_pag):
         return {
             "nfe40_indPag": ind_pag,
@@ -296,6 +307,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: infAdic
     ##########################
+
     nfe40_infAdFisco = fields.Char(
         compute="_compute_nfe40_additional_data",
     )
@@ -303,6 +315,7 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: infCpl
     ##########################
+
     nfe40_infCpl = fields.Char(
         compute="_compute_nfe40_additional_data",
     )
@@ -332,161 +345,14 @@ class NFe(spec_models.StackedModel):
     ##########################
     # NF-e tag: infRespTec
     ##########################
+
     nfe40_infRespTec = fields.Many2one(
         comodel_name="res.partner", related="company_id.technical_support_id"
     )
 
-    @api.multi
-    def _document_number(self):
-        # TODO: Criar campos no fiscal para codigo aleatorio e digito verificador,
-        # pois outros modelos também precisam dessescampos: CT-e, MDF-e etc
-        super()._document_number()
-        for record in self.filtered(filter_processador_edoc_nfe):
-            if record.document_key:
-                chave = ChaveEdoc(record.document_key)
-                record.nfe40_cNF = chave.codigo_aleatorio
-                record.nfe40_cDV = chave.digito_verificador
-
-    def _serialize(self, edocs):
-        edocs = super()._serialize(edocs)
-        for record in self.with_context({"lang": "pt_BR"}).filtered(
-            filter_processador_edoc_nfe
-        ):
-            inf_nfe = record.export_ds()[0]
-
-            tnfe = leiauteNFe.TNFe(infNFe=inf_nfe, infNFeSupl=None, Signature=None)
-            tnfe.original_tagname_ = "NFe"
-
-            edocs.append(tnfe)
-
-        return edocs
-
-    def _processador(self):
-        if not self.company_id.certificate_nfe_id:
-            raise UserError(_("Certificado não encontrado"))
-
-        certificado = cert.Certificado(
-            arquivo=self.company_id.certificate_nfe_id.file,
-            senha=self.company_id.certificate_nfe_id.password,
-        )
-        session = Session()
-        session.verify = False
-        transmissao = TransmissaoSOAP(certificado, session)
-        return edoc_nfe(
-            transmissao,
-            self.company_id.state_id.ibge_code,
-            versao=self.nfe_version,
-            ambiente=self.nfe_environment,
-        )
-
-    @api.multi
-    def _document_export(self, pretty_print=True):
-        super()._document_export()
-        for record in self.filtered(filter_processador_edoc_nfe):
-            record._export_fields_pagamentos()
-            edoc = record.serialize()[0]
-            processador = record._processador()
-            xml_file = processador._generateds_to_string_etree(
-                edoc, pretty_print=pretty_print
-            )[0]
-            _logger.debug(xml_file)
-            event_id = self.event_ids.create_event_save_xml(
-                company_id=self.company_id,
-                environment=(
-                    EVENT_ENV_PROD if self.nfe_environment == "1" else EVENT_ENV_HML
-                ),
-                event_type="0",
-                xml_file=xml_file,
-                document_id=self,
-            )
-            record.authorization_event_id = event_id
-
-    def atualiza_status_nfe(self, infProt, xml_file):
-        self.ensure_one()
-        # TODO: Verificar a consulta de notas
-        # if not infProt.chNFe == self.key:
-        #     self = self.search([
-        #         ('key', '=', infProt.chNFe)
-        #     ])
-        if infProt.cStat in AUTORIZADO:
-            state = SITUACAO_EDOC_AUTORIZADA
-        elif infProt.cStat in DENEGADO:
-            state = SITUACAO_EDOC_DENEGADA
-        else:
-            state = SITUACAO_EDOC_REJEITADA
-        if self.authorization_event_id and infProt.nProt:
-            if type(infProt.dhRecbto) == datetime:
-                protocol_date = fields.Datetime.to_string(infProt.dhRecbto)
-            else:
-                protocol_date = fields.Datetime.to_string(
-                    datetime.fromisoformat(infProt.dhRecbto)
-                )
-
-            self.authorization_event_id.set_done(
-                status_code=infProt.cStat,
-                response=infProt.xMotivo,
-                protocol_date=protocol_date,
-                protocol_number=infProt.nProt,
-                file_response_xml=xml_file,
-            )
-        self.write(
-            {
-                "status_code": infProt.cStat,
-                "status_name": infProt.xMotivo,
-            }
-        )
-        self._change_state(state)
-
-    @api.multi
-    def _eletronic_document_send(self):
-        super(NFe, self)._eletronic_document_send()
-        for record in self.filtered(filter_processador_edoc_nfe):
-            record._export_fields_pagamentos()
-            processador = record._processador()
-            for edoc in record.serialize():
-                processo = None
-                for p in processador.processar_documento(edoc):
-                    processo = p
-                    if processo.webservice == "nfeAutorizacaoLote":
-                        record.authorization_event_id._save_event_file(
-                            processo.envio_xml.decode("utf-8"), "xml"
-                        )
-
-            if processo.resposta.cStat in LOTE_PROCESSADO + ["100"]:
-                record.atualiza_status_nfe(
-                    processo.protocolo.infProt, processo.processo_xml.decode("utf-8")
-                )
-                if processo.protocolo.infProt.cStat in AUTORIZADO:
-                    try:
-                        record.make_pdf()
-                    except Exception as e:
-                        # Não devemos interromper o fluxo
-                        # E dar rollback em um documento
-                        # autorizado, podendo perder dados.
-
-                        # Se der problema que apareça quando
-                        # o usuário clicar no gera PDF novamente.
-                        _logger.error("DANFE Error \n {}".format(e))
-
-            elif processo.resposta.cStat == "225":
-                state = SITUACAO_EDOC_REJEITADA
-
-                self._change_state(state)
-
-                self.write(
-                    {
-                        "status_code": processo.resposta.cStat,
-                        "status_name": processo.resposta.xMotivo,
-                    }
-                )
-        return
-
-    @api.multi
-    def _document_date(self):
-        super()._document_date()
-        for record in self.filtered(filter_processador_edoc_nfe):
-            if not record.date_in_out:
-                record.date_in_out = fields.Datetime.now()
+    ################################
+    # Framework Spec model's methods
+    ################################
 
     def _export_field(self, xsd_field, class_obj, member_spec):
         if xsd_field == "nfe40_tpAmb":
@@ -503,7 +369,7 @@ class NFe(spec_models.StackedModel):
             self.nfe40_detPag and self.nfe40_detPag[0].nfe40_tPag == "90"
         ):
             return False
-        return super(NFe, self)._export_field(xsd_field, class_obj, member_spec)
+        return super()._export_field(xsd_field, class_obj, member_spec)
 
     def _export_many2one(self, field_name, xsd_required, class_obj=None):
         self.ensure_one()
@@ -527,10 +393,10 @@ class NFe(spec_models.StackedModel):
                 ):
                     return False
 
-        return super(NFe, self)._export_many2one(field_name, xsd_required, class_obj)
+        return super()._export_many2one(field_name, xsd_required, class_obj)
 
     def _export_one2many(self, field_name, class_obj=None):
-        res = super(NFe, self)._export_one2many(field_name, class_obj)
+        res = super()._export_one2many(field_name, class_obj)
         i = 0
         for field_data in res:
             i += 1
@@ -578,176 +444,3 @@ class NFe(spec_models.StackedModel):
             vals.update(new_value)
         else:
             super()._build_many2one(comodel, vals, new_value, key, value, path)
-
-    def view_pdf(self):
-        if not self.filtered(filter_processador_edoc_nfe):
-            return super().view_pdf()
-        if not self.authorization_file_id or not self.file_report_id:
-            self.make_pdf()
-        return self._target_new_tab(self.file_report_id)
-
-    def make_pdf(self):
-        if not self.filtered(filter_processador_edoc_nfe):
-            return super().make_pdf()
-
-        file_pdf = self.file_report_id
-        self.file_report_id = False
-        file_pdf.unlink()
-
-        if self.authorization_file_id:
-            arquivo = self.authorization_file_id
-            xml_string = base64.b64decode(arquivo.datas).decode()
-        else:
-            arquivo = self.send_file_id
-            xml_string = base64.b64decode(arquivo.datas).decode()
-            xml_string = self.temp_xml_autorizacao(xml_string)
-
-        pdf = base.ImprimirXml.imprimir(
-            string_xml=xml_string,
-            # output_dir=self.authorization_event_id.file_path
-        )
-        # TODO: Alterar a opção output_dir para devolter também o arquivo do XML
-        # no retorno, evitando a releitura do arquivo.
-
-        self.file_report_id = self.env["ir.attachment"].create(
-            {
-                "name": self.document_key + ".pdf",
-                "datas_fname": self.document_key + ".pdf",
-                "res_model": self._name,
-                "res_id": self.id,
-                "datas": base64.b64encode(pdf),
-                "mimetype": "application/pdf",
-                "type": "binary",
-            }
-        )
-
-    def temp_xml_autorizacao(self, xml_string):
-        """ TODO: Migrate-me to erpbrasil.edoc.pdf ASAP"""
-        root = etree.fromstring(xml_string)
-        ns = {None: "http://www.portalfiscal.inf.br/nfe"}
-        new_root = etree.Element("nfeProc", nsmap=ns)
-
-        protNFe_node = etree.Element("protNFe")
-        infProt = etree.SubElement(protNFe_node, "infProt")
-        etree.SubElement(infProt, "tpAmb").text = "2"
-        etree.SubElement(infProt, "verAplic").text = ""
-        etree.SubElement(infProt, "dhRecbto").text = None
-        etree.SubElement(infProt, "nProt").text = ""
-        etree.SubElement(infProt, "digVal").text = ""
-        etree.SubElement(infProt, "cStat").text = ""
-        etree.SubElement(infProt, "xMotivo").text = ""
-
-        new_root.append(root)
-        new_root.append(protNFe_node)
-        return etree.tostring(new_root)
-
-    def _document_cancel(self, justificative):
-        super()._document_cancel(justificative)
-        online_event = self.filtered(filter_processador_edoc_nfe)
-        if online_event:
-            online_event._nfe_cancel()
-
-    def _nfe_cancel(self):
-        self.ensure_one()
-        processador = self._processador()
-
-        if not self.authorization_protocol:
-            raise UserError(_("Authorization Protocol Not Found!"))
-
-        evento = processador.cancela_documento(
-            chave=self.document_key[3:],
-            protocolo_autorizacao=self.authorization_protocol,
-            justificativa=self.cancel_reason.replace("\n", "\\n"),
-        )
-        processo = processador.enviar_lote_evento(lista_eventos=[evento])
-        # Gravamos o arquivo no disco e no filestore ASAP.
-
-        self.cancel_event_id = self.event_ids.create_event_save_xml(
-            company_id=self.company_id,
-            environment=(
-                EVENT_ENV_PROD if self.nfe_environment == "1" else EVENT_ENV_HML
-            ),
-            event_type="2",
-            xml_file=processo.envio_xml.decode("utf-8"),
-            document_id=self,
-        )
-
-        for retevento in processo.resposta.retEvento:
-            if not retevento.infEvento.chNFe == self.document_key[3:]:
-                continue
-
-            if retevento.infEvento.cStat not in CANCELADO:
-                mensagem = "Erro no cancelamento"
-                mensagem += "\nCódigo: " + retevento.infEvento.cStat
-                mensagem += "\nMotivo: " + retevento.infEvento.xMotivo
-                raise UserError(mensagem)
-
-            if retevento.infEvento.cStat == CANCELADO_FORA_PRAZO:
-                self.state_fiscal = SITUACAO_FISCAL_CANCELADO_EXTEMPORANEO
-            elif retevento.infEvento.cStat == CANCELADO_DENTRO_PRAZO:
-                self.state_fiscal = SITUACAO_FISCAL_CANCELADO
-
-            self.state_edoc = SITUACAO_EDOC_CANCELADA
-            self.cancel_event_id.set_done(
-                status_code=retevento.infEvento.cStat,
-                response=retevento.infEvento.xMotivo,
-                protocol_date=fields.Datetime.to_string(
-                    datetime.fromisoformat(retevento.infEvento.dhRegEvento)
-                ),
-                protocol_number=retevento.infEvento.nProt,
-                file_response_xml=processo.retorno.content.decode("utf-8"),
-            )
-
-    def _document_correction(self, justificative):
-        super()._document_correction(justificative)
-        online_event = self.filtered(filter_processador_edoc_nfe)
-        if online_event:
-            online_event._nfe_correction(justificative)
-
-    def _nfe_correction(self, justificative):
-        self.ensure_one()
-        processador = self._processador()
-
-        numeros = self.event_ids.filtered(
-            lambda e: e.type == "14" and e.state == "done"
-        ).mapped("sequence")
-
-        sequence = str(int(max(numeros)) + 1) if numeros else "1"
-
-        evento = processador.carta_correcao(
-            chave=self.document_key[3:],
-            sequencia=sequence,
-            justificativa=justificative.replace("\n", "\\n"),
-        )
-        processo = processador.enviar_lote_evento(lista_eventos=[evento])
-        # Gravamos o arquivo no disco e no filestore ASAP.
-        event_id = self.event_ids.create_event_save_xml(
-            company_id=self.company_id,
-            environment=(
-                EVENT_ENV_PROD if self.nfe_environment == "1" else EVENT_ENV_HML
-            ),
-            event_type="14",
-            xml_file=processo.envio_xml.decode("utf-8"),
-            document_id=self,
-            sequence=sequence,
-            justification=justificative,
-        )
-        for retevento in processo.resposta.retEvento:
-            if not retevento.infEvento.chNFe == self.document_key[3:]:
-                continue
-
-            if retevento.infEvento.cStat not in EVENTO_RECEBIDO:
-                mensagem = "Erro na carta de correção"
-                mensagem += "\nCódigo: " + retevento.infEvento.cStat
-                mensagem += "\nMotivo: " + retevento.infEvento.xMotivo
-                raise UserError(mensagem)
-
-            event_id.set_done(
-                status_code=retevento.infEvento.cStat,
-                response=retevento.infEvento.xMotivo,
-                protocol_date=fields.Datetime.to_string(
-                    datetime.fromisoformat(retevento.infEvento.dhRegEvento)
-                ),
-                protocol_number=retevento.infEvento.nProt,
-                file_response_xml=processo.retorno.content.decode("utf-8"),
-            )
